// dashboard/src/utils/hasura-api/generated/schemas/trackLogicalModelStep.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { TrackLogicalModelArgs } from './trackLogicalModelArgs';
import type { TrackLogicalModelStepType } from './trackLogicalModelStepType';

export interface TrackLogicalModelStep {
  type: TrackLogicalModelStepType;
  args: TrackLogicalModelArgs;
}