// dashboard/src/utils/hasura-api/generated/schemas/untrackLogicalModelStep.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { UntrackLogicalModelArgs } from './untrackLogicalModelArgs';
import type { UntrackLogicalModelStepType } from './untrackLogicalModelStepType';

export interface UntrackLogicalModelStep {
  type: UntrackLogicalModelStepType;
  args: UntrackLogicalModelArgs;
}